Create types file and fix imports
