+++ src/state_manager.ts
@@ -1,16 +1,13 @@
 import { InteractionHooks, InterfacePlayer } from "./players/interface_player";
-import {
-  MintWorksEngine,
-  MintWorksEngineState,
-  MintWorksParams,
-  PlayerWithInformation,
-} from "./engine";
+import { MintWorksEngine, MintWorksEngineState, MintWorksParams } from "./engine";
+
 import { Neighbourhood } from "./neighbourhood";
 import { Turn } from "./turn";
 import { Building, HandPlan, Plan } from "./plan";
 import { LocationCard, createLocationsFromState } from "./location";
 import { addHooksToPlans } from "./plans";
 import { MintWorksTurnFactory } from "./turn_factory";
+import { PlayerWithInformation } from "./types";
 
 /**
  *
